requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
Flask-JWT-Extended==4.5.2
bcrypt==4.1.3
python-dotenv==1.0.0